fix <rdar://problem/19756953> Swift error: cannot capture '$0' before it is declared
This sets the location of the implicit closure decls (like $0) to being the location
of the { in a ClosureExpr, instead of the location of the first use.  The capture tracker
uses source location information of the decl and the DeclRefExpr to determine if the
referenced value was captured too early, which is what is causing this incorrect error.



Swift SVN r28802