Skip some tests at -Ounchecked, if they expect crashes.
In most cases when we are testing traps the right choice is to skip the test when we run with -Ounchecked, because some of the expected crashes might or might not cause a crash in -Ounchecked mode.

Discussed and agreed with Dmitri.

With this change and our recent bug-fixing efforts, running the whole test-suite + validation-suite with -Ounchecked results in only one failing test.

Swift SVN r28806