IRGen: get unchecked_ref_cast working with class existential operand
There was some duplicated code and we weren't doing a claimAll() on the
input explosion in one of the duplicates.

This partially reverts r28686, but keeping the test and updating it.

Second version of this patch makes the SIL verifier and combiner passes
use the same predicate, fixing assertion failures when casting
metatypes.

Fixes <rdar://problem/21005110>.

Swift SVN r28873