stdlib: Tweak how empty enum cases are printed
Previously a mirror for an empty case would have one entry, consisting
of the case name and a value of ().

Now, the mirror's summary shows both the type and the case.  If there is
no payload, the mirror will not have any children. The presence of a
child indicates there is an actual payload.

Also put a non-ASCII character in the reflection test to ensure the
runtime's primitives can round-trip UTF8.

Fixes <rdar://problem/20994093>.

Swift SVN r28874