Warn about non-@objc potential witnesses for optional requirements.
When an optional requirement goes unfulfilled by a conformance, and
there is a non-@objc declaration with that name in the type/extension
declaring conformance, warn that it does not satisfy the optional
requirement. Sadly, this diagnostic is long because there are notes
for the two potential fixes: add @objc to try to conform, or move the
declaration elsewhere to silence the compiler.

Fixes rdar://problem/20219297.

Swift SVN r28908