On arm we flush subnormals to zero so the subnormal test will not work.
To be a bit more precise on arm we use neon instructions for floating point
operations which flush to zero.

When we optimize this code we use a multiply by inverse instead of a division.
The multiply is instruction selected to a a neon instruction and we get zero
instead of a denormal.

Swift SVN r28913