Remove asserts that cause compilation failures on Linux.
The Linux assert() is apparently doing a positive test on the value,
whereas the Darwin assert() is negating it.

APInt has an operator!(), but APInt cannot be implicitly converted to a
bool, so directly testing the value caused issues only on the Linux
build.

JoeG says the asserts were testing pointer nullness before his recent
representation changes, and aren't meaningful with the new
representation, so they should just be removed.

Swift SVN r29011