QoI: Improve diagnostic when protocol witness is less available than protocol requires.
Instead of the rather inscrutable "'foo()' is less available than protocol requires" be
more helpful: "protocol 'HasFoo' requires 'foo()' to be available on OS X 10.9 and newer".

Also add a note indicating where the conformance was introduced, as this may be in a
separate location from both the witness and requirement.

rdar://problem/20610411

Swift SVN r29028