[CodeCompletion] Avoid assert when top-level if contains an autoclosure
We have an unreachable in the visitor for auto closures, so avoid
visiting it.  It would be great to have a clearer picture of what needs
to happen for AutoClosure vs ClosureExpr. It's hard to tell exactly
where AutoClosure is supposed to be handled if at all.

Swift SVN r29037