Disable argument explosion in function signature optimization.
We currently eagerly explode arguments during function signature
optimization, which gets in the way of the IRGen-level decision of
whether to pass them indirectly.

This disables the explosion in the optimization pass until we can go
back and implement it in a way that we expect it to be more beneficial
than the indirect argument passing (rdar://problem/21114206).

It looks like CaptureProp and possibly a couple other benchmarks
regress, which I'll confirm and open radars for.

Resolves rdar://problem/21093134.

Swift SVN r29047