Introduce a ValueLifetime utility and migrate some clients.
I needed this for nontrivial DeadObjectElimination.

This also makes the order that we remove uses and insert releases
deterministic (although order of removal should not be required).

The utility only does what it needs to do, nothing more. The result is
separate from the Analysis. ValueLifetimeAnalysis does not impose any
structure on the client. The client decides how to transform or filter
the user list. The client that uses the Analysis can access its
internal book-keeping and helpers.

I should be able to migrate ReleaseTracker easily.

Swift SVN r29165