[sil-devirtualizer] Fix a bug related to devirtualization of bound generic classes.
ClassHierarchyAnalysis reports relationships between classes as they are all unbound. Therefore, when we try to devirtualize a method of a bound generic class, we need to exclude non-subclasses from the list of subclasses reported by ClassHierarchyAnalysis.

This fixes a compiler crash on generic_class.swift test when it is built with -O.

rdar://21188939.

Swift SVN r29215