Do not generate unchecked_ref_bit_cast from unchecked_enum_data.
We were generating unchecked_ref_bit_cast in cases where we had an
unchecked_enum_data with an unchecked_ref_bit_cast as an argument and
we're accessing the first case with a payload. We don't know anything
about the sizes of the types involved, and as a result at IRGen-time we
may end up generating a trap if the bit cast is casting between
differently-sized types.

We really shouldn't be making any assumptions about layout at the SIL
level, at least about arbitrary types.

There is at least one other similar transformation that I will take a
look at to see if it could be problematic as well.

Fixes rdar://problem/21202769.

Swift SVN r29311