Fix Objective-C encodings for Swift-defined @objc classes and protocols.
We use Clang to produce Objective-C type encodings (e.g., for property
descriptors). In cases where we need to synthesize a Clang @interface
or @protocol to map a Swift type into a Clang type for this purpose,
attach an objc_runtime_name attribute to the synthesized Clang
declaration so that it gets the appropriate Objective-C runtime name
in the encoding, rather than simply falling back to the (unmangled)
Swift class or protocol name.

This requires a paired Clang commit, since Clang was ignoring the
objc_runtime_name attribute in its type encoding.

Swift SVN r29408