[IRGen] Bitcast SIL globals that don't have the right LLVM type.
This comes up when the SIL global actually refers to a static global defined
in C. In these cases Clang may have already emitted the global with a
different type---one appropriate for Clang, but not for Swift. By the time
we get to this part of the code, we should have already rejected anything not
layout-compatible, so just do the bitcast and be done with it.

Reviewed by John.

rdar://problem/21361469

Swift SVN r29471