[ls-opts] Properly distinguish in between Stores with read dependencies which can not be dead but still can be forwarded from and those stores which can still be dead.
Previously if we saw a read dependency, we would stop tracking the load for
forwarding purposes even though we were really just trying to say that the load
could no longer be dead.

Additionally, we were not properly setting read dependency in the case where we
had a pointer alias, but not an exact match. This would have resulted in certain
stores being considered dead incorrectly.

I also added a bunch of validation code to maintain invariants and a bunch of tests.

rdar://21376211

Swift SVN r29516