Add in a DeletedInstSet to SILCombine to prevent dangling deleted instructions from being added to the Worklist.
Specifically, when inserting instructions from the tracking list into
the Worklist we first check if the instruction is in the
DeletedInstSet. If the instruction is, we do not add it to the worklist.

This bug comes up if one adds an instruction using a SILBuilder and then
deletes it in the same iteration of SILCombine. I noticed it when debug
print outs were trying to access the parent of the dangling deleted
instruction causing the program to blow up.

Swift SVN r29521