SILPasses: Teach DI about delegating constructors
Even when we're constructing a reference type, a delegating constructor
call might still return self via an out parameter and involve assignments,
in particular if we're delegating to a constructor defined in an
extension.

Note that an actual assignment to self is prohibited in reference type
constructors, but if we see one of those from Sema we have a bug.

Fixes <rdar://problem/21475240>.

Swift SVN r29556