Bump iteration limit in inliner.
This counter is in place to defend against any chance we go into an
infinite loop due to failing to recognize recursive inlining. In debug
builds we assert if we hit the limit. In release builds we bail
out.

Since we shipped with asserts enabled, a user has reported hitting the
limit (and associated assert). In this particular case we're not going
into an infinite loop, and bumping the count allows their code to
compile.

While I continue to look at the test (in rdar://problem/21391092) case I
want to make sure we're not blocking people unnecessarily, so I'm
bumping the limit.

Swift SVN r29575