Drop the correct 'Self' constraint for a protocol member.
We were dropping *all* constraints on the Self type when forming
constraints from a protocol member, which included the extra
constraints provided by constrained extensions. Only drop the actual
protocol in which the member occurs (or the protocol extended by the
containing extension).

Fixes rdar://problem/21401180 and the 8 dupes I've found so far.

Swift SVN r29708