two changes here:  - When diagnosing an error passing a noescape function to an escapeing function pointer,    return 'true' to avoid follow-on bogus error messages and notes being produced.  - Start classifying the closeness of overload matches, to allow future diagnostics to be    improved.
Note that the second regresses the testcase in test/Constraints/generics.swift because the
decomposeArgumentType function doesn't know to look through the archetype for the protocol
present on the operator, and thus thinks the function only takes one argument.  Advice on
how to best detect this situation is appreciated.



Swift SVN r29731